UPDATE DELETE
saat delete maka gambar juga terhapus difolder

+++ api/controllers/desawisata.js
@@ -109,7 +109,7 @@ exports.updateDesaWisata = async (req, res) => {
       desaWisata.nama = nama;
     }
     if (gambar) {
-     // Cek apakah ada gambar sebelumnya, jika ada maka hapus gamabr sebelumnya
+     // Cek apakah ada gambar sebelumnya, jika ada maka hapus gambar sebelumnya
       if (desaWisata.gambar) {
         const imagePath = path.join(
           __dirname,
@@ -171,12 +171,29 @@ exports.updateDesaWisata = async (req, res) => {
 exports.deleteDesaWisata = async (req, res) => {
     try {
       const { id } = req.params;
+      const desawisata = await DesaWisata.findOne({ where: { id: id} });
+
+      if (!desawisata) {
+        return res.status(404).json({ message: "Desa Wisata tidak ditemukan" });
+      }
+  
+      const gambarDesawisataLama = desawisata.gambar;
+
       const deletedDesawisataCount = await DesaWisata.destroy({
         where: {
           id: id,
         },
       });
       if (deletedDesawisataCount > 0) {
+        if (gambarDesawisataLama) {
+            const imagePath = path.join(__dirname, "../../uploads/resource/desawisata", gambarDesawisataLama);
+            if (fs.existsSync(imagePath)) {
+              fs.unlinkSync(imagePath);
+              console.log("Foto Desa berhasil dihapus:", gambarDesawisataLama);
+            } else {
+              console.log("Foto Desa tidak ditemukan:", gambarDesawisataLama);
+            }
+          }
         return res.status(200).json({ message: "Desa wisata berhasil dihapus" });
       } else {
         return res.status(404).json({ message: "Desa wisata tidak ditemukan" });
